Added wrapper from mzp_t and integrated it in zobrist

+++ Engine/IHashAlgorithm.h
@@ -1,7 +1,9 @@
 #include <string>
+#include <gmp.h>
 #include "Move.h"
 #include "Cell.h"
 #include "IBoard.h"
+#include "BigInteger.h"
 
 namespace engine{
 
@@ -13,7 +15,7 @@ namespace engine{
 		}
 
 		virtual void UpdateBoard(IBoard* board) = 0;
-		virtual long long GetHash() = 0;
+		virtual BigInteger GetHash() = 0;
 
 	};
 }
