+++ Engine/TranspositionTable.h
@@ -1,24 +1,25 @@
 #pragma once
 #include <map>
 #include <queue>
-#include <iostream>
+#include <gmp.h>
 #include "TranspositionTableData.h"
 #include "Move.h"
 #include "Players.h"
+#include "BigInteger.h"
 using namespace std;
 
 namespace engine {
 	class ENGINE_API TranspositionTable
 	{
 	private:
-		map<long long, TranspositionTableData*>* _hashMap;
-		queue<long long>* _ageQueue;
+		map<BigInteger, TranspositionTableData*>* _hashMap;
+		queue<BigInteger>* _ageQueue;
 		int _maxSize;
 	public:
 		TranspositionTable(int maxSize);
 		~TranspositionTable(void);
-		void Insert(long long key, int depth, int score, Players player, Move* bestMove);
-		bool Contains(long long key);
-		TranspositionTableData* Get(long long key);
+		void Insert(BigInteger key, int depth, int score, Players player, Move* bestMove);
+		bool Contains(BigInteger key);
+		TranspositionTableData* Get(BigInteger key);
 	};
 }
